requirements.txt
# StockAdvisor Backend - Python Dependencies
# Created by Digital COE Gen AI Team

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
starlette==0.35.1

# Database
motor==3.3.2
pymongo==4.6.1
beanie==1.25.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# WebSocket
websockets==12.0
python-socketio==5.10.0

# AI/ML
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0

# Stock Data APIs
yfinance==0.2.35
alpha_vantage==2.3.1

# Caching
redis==5.0.1
aiocache==0.12.2

# Task Queue
celery==5.3.6
flower==2.0.1

# Logging & Monitoring
loguru==0.7.2
sentry-sdk[fastapi]==1.39.1
prometheus-fastapi-instrumentator==6.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Utils
python-dotenv==1.0.0
orjson==3.9.10

